Refactor: Extract interface for PNC
