+++ communication/src/main/java/org/jboss/da/communication/pnc/impl/PNCConnectorImpl.java
@@ -1,5 +1,6 @@
-package org.jboss.da.communication.pnc;
+package org.jboss.da.communication.pnc.impl;
 
+import org.jboss.da.communication.pnc.api.PNCConnector;
 import org.jboss.da.common.util.Configuration;
 import org.jboss.da.common.util.ConfigurationParseException;
 import org.jboss.da.communication.pnc.authentication.PNCAuthentication;
@@ -14,13 +15,13 @@ import javax.ws.rs.core.MediaType;
 import java.util.Arrays;
 import java.util.List;
 
-public class PNC {
+public class PNCConnectorImpl implements PNCConnector {
 
     private Configuration config;
     private String pncServer;
     private PNCAuthentication pncAuthenticate;
 
-    public PNC() throws ConfigurationParseException {
+    public PNCConnectorImpl() throws ConfigurationParseException {
         config = new Configuration();
         pncServer = config.getConfig().getPncServer();
         pncAuthenticate = new PNCAuthentication();
@@ -44,21 +45,25 @@ public class PNC {
         return getClient(endpoint, false);
     }
 
+    @Override
     public List<BuildConfiguration> getBuildConfigurations() throws Exception {
         ClientResponse<BuildConfiguration[]> response = getClient("build-configurations").get(BuildConfiguration[].class);
         return Arrays.asList(response.getEntity());
     }
 
+    @Override
     public List<BuildConfigurationSet> getBuildConfigurationSets() throws Exception {
         ClientResponse<BuildConfigurationSet[]> response = getClient("build-configuration-sets").get(BuildConfigurationSet[].class);
         return Arrays.asList(response.getEntity());
     }
 
+    @Override
     public List<Product> getProducts() throws Exception {
         ClientResponse<Product[]> response = getClient("products").get(Product[].class);
         return Arrays.asList(response.getEntity());
     }
 
+    @Override
     public List<Project> getProjects() throws Exception {
         ClientResponse<Project[]> response = getClient("projects").get(Project[].class);
         return Arrays.asList(response.getEntity());
